ai-service/requirements.txt
# AI 智能打标服务依赖

# Web 框架
fastapi==0.104.1          # 高性能 Web 框架
uvicorn[standard]==0.24.0 # ASGI 服务器

# 深度学习
tensorflow==2.14.0        # TensorFlow 深度学习框架
# 如果有 NVIDIA GPU，可以使用 tensorflow-gpu==2.14.0

# 图像处理
Pillow==10.1.0           # Python 图像处理库

# HTTP 客户端
requests==2.31.0         # HTTP 请求库

# 数据验证
pydantic==2.5.0          # 数据验证和设置管理

# 工具库
numpy==1.24.3            # 数值计算库
python-dotenv==1.0.0     # 环境变量管理
tqdm==4.66.1             # 进度条显示

# 可选：如果需要连接 RabbitMQ
# pika==1.3.2            # RabbitMQ 客户端

# 可选：如果需要连接 Nacos
# nacos-sdk-python==0.1.7  # Nacos 服务注册与发现

# 开发工具（可选）
# pytest==7.4.3          # 测试框架
# black==23.11.0         # 代码格式化
# flake8==6.1.0          # 代码检查
